fixes and improvements
1- add shadow Root to isolate contents from website
2- fix multiple slashs issue in iframe
