+++ src/modes/index.js
@@ -1,0 +1,12 @@
+import { shortUrlMode } from './shortURL';
+
+export const modes = {
+  /**
+       * @param {import('./index').default} that
+       */
+  'short-url': shortUrlMode
+  /**
+    * TODO: Implement mode for transferID,
+    * where the transferID is used to retrieve the iframe url from the interaction API
+    */
+};
