+++ src/utils.js
@@ -1,133 +1,6 @@
-import { shortURLPublic } from 'videoengager-js-sdk';
-export function checkValue (value = '') {
-  if (!value) {
-    return false;
-  }
-  let val = value.trim();
-  // strip the value from any query params or other stuff
-  if (val.includes('/')) {
-    val = val.split('/')[0];
-  }
-  if (val.includes('?')) {
-    val = val.split('?')[0];
-  }
-  if (val.includes('&')) {
-    val = val.split('&')[0];
-  }
-  if (val.includes('=')) {
-    val = val.split('=')[0];
-  }
-  if (val.includes('#')) {
-    val = val.split('#')[0];
-  }
-  if (!val) {
-    return false;
-  }
-  return val;
-}
+import { modes } from './modes';
 
-export function checkIfPathIncludesCodeOrPin (fullPath) {
-  try {
-    let value = null;
-    if (fullPath.includes('/ve/')) {
-      value = checkValue(fullPath.split('/ve/')[1]);
-      if (!value) {
-        throw new Error('Invalid Code value');
-      }
-      return {
-        type: 'code',
-        value: value
-      };
-    } else if (fullPath.includes('/pin/')) {
-      value = checkValue(fullPath.split('/pin/')[1]);
-      if (!value) {
-        throw new Error('Invalid Pin value');
-      }
-      return {
-        type: 'pin',
-        value: value
-      };
-    }
-    console.error('Invalid path: ', fullPath, 'Please use /ve/ or /pin/');
-    return null;
-  } catch (error) {
-    console.error(error);
-    return null;
-  }
-}
-export async function getIframeUrlFromShortCodeOrPin ({ code, pin }, tenantID, serverUrl = 'https://videoengager.com') {
-  const data = {
-    iframeUrl: null,
-    error: true
-  };
-  if (code) {
-    const response = await shortURLPublic.getShortURLByCode({
-      tenantID,
-      code
-    }).catch((error) => {
-      console.error(error);
-      data.error = true;
-    });
-    if (response?.data?.active) {
-      data.iframeUrl = `${serverUrl}/${response.data.url}`;
-    } else {
-      data.error = true;
-    }
-  } else if (pin) {
-    const response = await shortURLPublic.getShortURLByPin({
-      tenantID,
-      pin
-    }).catch((error) => {
-      console.error(error);
-      data.error = true;
-    });
-    if (response?.data?.shortUrl) {
-      data.iframeUrl = `${serverUrl}/${response.data.shortUrl}`;
-    }
-    // Handle PIN case here
-  }
-
-  return data;
-}
-export const modes = {
-  /**
-     * @param {import('./index').default} that
-     */
-  'short-url': async function shortUrl (that) {
-    try {
-      const fullPath = window.location.href;
-      const isReady = that.isReadyToUseVEAPI;
-      if (!isReady) {
-        console.log('isReady', isReady);
-        console.error('VideoEngager API is not ready yet');
-        return false;
-      }
-      const data = checkIfPathIncludesCodeOrPin(fullPath);
-      if (!data) {
-        console.error('Invalid path');
-        throw new Error('(ERROR: Invalid path)');
-      }
-      const { type, value } = data;
-      const tenantID = that.getAttribute('ve-tenant');
-      const serverUrl = that.getAttribute('ve-server-url');
-      const iframeUrlData = await getIframeUrlFromShortCodeOrPin({ [type]: value }, tenantID, serverUrl);
-      if (iframeUrlData.iframeUrl) {
-        that.createIframe(iframeUrlData.iframeUrl);
-      } else {
-        throw new Error('Cannot get iframe url from API');
-      }
-    } catch (error) {
-      console.error(error);
-      that.handleInvalidShortCodeOrPin(error.response?.data?.invalidUrl || false);
-    }
-    //   manual: 'manual'
-  }
-  /**
-  * TODO: Implement mode for transferID,
-  * where the transferID is used to retrieve the iframe url from the interaction API
-  */
-};
-export function checkIfUrlIsCorrect (url) {
+export function checkServerUrl (url) {
   if (!url) {
     return false;
   }
@@ -158,4 +31,17 @@ export const invalidOrExpiredBehavoirs = {
   'show-text-element': 'show-text-element',
   redirect: 'redirect',
   hide: 'hide'
+};
+export const attributesDescritions = {
+  've-server-url': 'The URL of the VideoEngager server (Required)',
+  've-tenant': 'The tenant ID (Required)',
+  've-widget-mode': 'The detection mode of videoCall, default (short-url) available options:' + Object.keys(modes).join(', '),
+  've-auto-start': 'Disable auto start, to start manually use method of element: element.startVideoEngagerCall()',
+  've-on-invalid-behavior': 'The behavior when the URL is invalid, default (redirect) available options:' + Object.keys(invalidOrExpiredBehavoirs).join(', '),
+  've-on-end-behavior': 'The behavior when the call is ended, default (redirect) available options:' + Object.keys(endOfCallsBehavoirs).join(', '),
+  've-on-invalid-text-element': 'The text to show when the URL is invalid or expired (only when ve-on-invalid-behavior="show-text-element")',
+  've-on-end-text-element': 'The text to show when the call is ended (only when ve-on-end-behavior="show-text-element")',
+  've-on-invalid-redirect-url': 'The URL to redirect to when the URL is invalid or expired (only when ve-on-invalid-behavior="redirect")',
+  've-on-end-redirect-url': 'The URL to redirect to when the call is ended (only when ve-on-end-behavior="redirect")',
+  've-loading-text-element': 'The text to show when the data is being loaded from the server (default: Loading...)'
 };
